add optional chaining in callback
