+++ README.md
@@ -20,6 +20,13 @@ yarn
 yarn start
 ```
 
+## 배포 사이트
+https://wordle-challenge-psi.vercel.app/
+
+## 구현영상 
+![](May-13-2022%2019-25-24.gif)
+
+
 ## Requirements 
 [x] 키보드를 클릭하여 입력받을 수 있다.   
 [x] 키보드를 클릭시 5개 이상은 입력되지 않는다.   
@@ -135,10 +142,9 @@ yarn start
     또한 기존의 useKeyboard의 결함이라고 생각했던 큰 파일 사이즈는 절대 큰 사이즈가 아니였다.
     그래서 useGame내부로 위치시켰고 기존의 useResetGame도 알맞은 위치로 옮길 수 있었다.
 
+--- 
 
-## 구현영상 
-![](May-13-2022%2019-25-24.gif)
-
+12) App.tsx의 구조에 대한 피드백, 없어도 되는 로직이 너무 많다??
 
 ## 이번 pr에서 받고 싶은 내용들 
 1) useKeyboard, useModalMessage hook의 경우 쓰임새가 적합한지 궁금하니다. 
@@ -166,3 +172,9 @@ yarn start
     v는 jsx, m과 c를 hook으로 분리하게 됩니다.
     model은 ajax로 받아온 데이터를 상태로 변화시킨 것 같고, c는 상태를 관리하는 비즈니스 로직이라 생각한다.
     이때 hook으로 분리를 하는 것은 상태와 상태를 변화시키는 동작들의 범위를 고립시켜서 재사용성과 독립성을 확보하기 위한 의미
+
+
+2. 룩업 테이블 :
+    다중 if 문의 경우 lookup 테이블을 이용하면 훨씬 수월하게 처리할 수 있다.
+    ![](2022-05-15-12-30-39.png)
+
